licensing: Change copyleft wording
should -> must. It is a legal requirement. We will probably talk about
the "should" aspect of it in the section on community, e.g. generally
rude to license your additions to a permissive license project under
copyleft.

"Harder to use and distribute". Additional requirements could be
described as "harder", but in most cases I don't think it's any
harder. Want to avoid pushing people away from a specific type of
license. If we want to spend time on it we could cover some details of
some of the harder parts of copyleft, but I don't want to just say
"harder". Of course we say more than waht is just on the slides, but I
feel like the slides shouldn't boil copyleft down to "harder".